ci: add translation rate to manifest
